Split CAN comms into source and header files
