+++ board/comms_constants.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define MAX_CAN_MSGS_PER_USB_BULK_TRANSFER 51U
+#define MAX_CAN_MSGS_PER_SPI_BULK_TRANSFER 170U
