+++ board/main_definitions.h
@@ -1,14 +1,3 @@
-#include "main_declarations.h"
+#pragma once
 
-// ********************* Globals **********************
-uint8_t hw_type = 0;
-board *current_board;
-uint32_t uptime_cnt = 0;
-
-// heartbeat state
-uint32_t heartbeat_counter = 0;
-bool heartbeat_lost = false;
-bool heartbeat_disabled = false;            // set over USB
-
-// siren state
-bool siren_enabled = false;
+#include "board/main_declarations.h"
